requirements.txt
asgiref==3.4.1
Django==3.2.10
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.8.3
Markdown==3.3.6
pkg-resources==0.0.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
zipp==3.6.0
